refactor: create split board page
