+++ frontend/src/components/CreateBoard/RegularBoard/SelectTeam/styles.tsx
