+++ frontend/src/components/CreateBoard/SplitBoard/SubTeamsTab/CreateBoardItem/CreateBoardItem.tsx
@@ -1,33 +1,29 @@
+import { deepClone } from 'fast-json-patch';
 import React from 'react';
 import { SetterOrUpdater } from 'recoil';
-import { deepClone } from 'fast-json-patch';
-
-import { highlight2Colors } from '@/styles/stitches/partials/colors/highlight2.colors';
-import { styled } from '@/styles/stitches/stitches.config';
 
 import LeftArrow from '@/components/CardBoard/CardBody/LeftArrow';
-import Icon from '@/components/Primitives/Icons/Icon/Icon';
 import Avatar from '@/components/Primitives/Avatars/Avatar/Avatar';
-import Box from '@/components/Primitives/Layout/Box/Box';
+import AvatarGroup from '@/components/Primitives/Avatars/AvatarGroup/AvatarGroup';
+import Icon from '@/components/Primitives/Icons/Icon/Icon';
 import Flex from '@/components/Primitives/Layout/Flex/Flex';
 import Separator from '@/components/Primitives/Separator/Separator';
 import Text from '@/components/Primitives/Text/Text';
 import { CreateBoardData } from '@/store/createBoard/atoms/create-board.atom';
+import { highlight2Colors } from '@/styles/stitches/partials/colors/highlight2.colors';
 import { BoardToAdd } from '@/types/board/board';
 import { BoardUserToAdd } from '@/types/board/board.user';
 import { BoardUserRoles } from '@/utils/enums/board.user.roles';
 import { getInitials } from '@/utils/getInitials';
-import AvatarGroup from '@/components/Primitives/Avatars/AvatarGroup/AvatarGroup';
+import { StyledBoardItem, WandButtonWrapper } from './styles';
 
-interface SubCardBoardProps {
+interface CreateBoardItemProps {
   index: number;
   board: BoardToAdd;
   setBoard: SetterOrUpdater<CreateBoardData>;
 }
 
-const Container = styled(Flex, Box, {});
-
-const SubCardBoard: React.FC<SubCardBoardProps> = ({ board, index, setBoard }) => {
+const CreateBoardItem: React.FC<CreateBoardItemProps> = ({ board, index, setBoard }) => {
   const { users } = board;
   const responsible = users.find((user) => user.role === BoardUserRoles.RESPONSIBLE)?.user;
 
@@ -70,91 +66,28 @@ const SubCardBoard: React.FC<SubCardBoardProps> = ({ board, index, setBoard }) =
   };
 
   return (
-    <Flex css={{ flex: '1 1 0', width: '100%' }}>
+    <Flex>
       <LeftArrow index={index} isDashboard={false} />
-
-      <Container
-        align="center"
-        elevation="1"
-        justify="between"
-        css={{
-          backgroundColor: 'white',
-          height: '$64',
-          width: '100%',
-          ml: '$40',
-          py: '$16',
-          pl: '$32',
-          pr: '$24',
-        }}
-      >
-        <Flex align="center" justify="start">
+      <StyledBoardItem align="center" elevation="1" justify="between">
+        <Flex css={{ flex: 1 }}>
           <Text heading="5">{board.title}</Text>
         </Flex>
-
-        <Flex align="center" justify="start" css={{ width: '50%' }}>
-          <Flex align="center" css={{ minWidth: '$160' }}>
-            <Text css={{ mr: '$8' }}>Responsible Lottery</Text>
+        <Flex align="center" css={{ flex: 3 }}>
+          <Flex align="center" gap={8}>
+            <Text>Responsible Lottery</Text>
             <Separator orientation="vertical" size="md" />
-          </Flex>
-
-          <Flex align="center" css={{ minWidth: 0 }}>
-            {users.length <= 1 ? (
-              <Flex
-                align="center"
-                justify="center"
-                css={{
-                  height: '$24',
-                  minWidth: '$24',
-                  width: '$24',
-                  borderRadius: '$round',
-                  border: '1px solid $colors$primary400',
-                  ml: '$12',
-                  opacity: '0.2',
-                }}
-                onClick={handleLottery}
-              >
-                <Icon
-                  name="wand"
-                  css={{
-                    width: '$12',
-                    height: '$12',
-                  }}
-                />
-              </Flex>
-            ) : (
-              <Flex
-                align="center"
-                justify="center"
-                css={{
-                  height: '$24',
-                  minWidth: '$24',
-                  width: '$24',
-                  borderRadius: '$round',
-                  border: '1px solid $colors$primary400',
-                  ml: '$12',
-                  cursor: 'pointer',
-                  transition: 'all 0.2s ease-in-out',
-                  '&:hover': {
-                    backgroundColor: '$primary400',
-                    color: 'white',
-                  },
-                }}
-                onClick={handleLottery}
-              >
-                <Icon
-                  name="wand"
-                  css={{
-                    width: '$12',
-                    height: '$12',
-                  }}
-                />
-              </Flex>
-            )}
-            <Flex css={{ minWidth: 0 }}>
+            <WandButtonWrapper
+              align="center"
+              justify="center"
+              disabled={users.length <= 1}
+              {...(users.length > 1 && { onClick: handleLottery })}
+            >
+              <Icon name="wand" size={12} />
+            </WandButtonWrapper>
+            <Flex>
               <Text
                 color="primary300"
                 css={{
-                  mx: '$8',
                   textOverflow: 'ellipsis',
                   whiteSpace: 'nowrap',
                   overflow: 'hidden',
@@ -165,7 +98,7 @@ const SubCardBoard: React.FC<SubCardBoardProps> = ({ board, index, setBoard }) =
               </Text>
             </Flex>
             <Avatar
-              css={{ position: 'relative', minWidth: '$34' }}
+              css={{ position: 'relative' }}
               fallbackText={getInitials(
                 responsible?.firstName ?? '-',
                 responsible?.lastName ?? '-',
@@ -178,14 +111,13 @@ const SubCardBoard: React.FC<SubCardBoardProps> = ({ board, index, setBoard }) =
             />
           </Flex>
         </Flex>
-
-        <Flex align="center" gap="8" justify="center">
+        <Flex align="center" justify="end" gap="8" css={{ flex: 2 }}>
           <Text size="sm">Sub team {index + 1}</Text>
           <AvatarGroup listUsers={board.users} userId="1" hasDrawer />
         </Flex>
-      </Container>
+      </StyledBoardItem>
     </Flex>
   );
 };
 
-export default SubCardBoard;
+export default CreateBoardItem;
